requirement.txt
asgiref==3.5.2
cffi==1.15.0
Django==4.0.4
gevent==21.12.0
greenlet==1.1.2
pycparser==2.21
pyzmq==23.0.0
sqlparse==0.4.2
tzdata==2022.1
zope.event==4.5.0
zope.interface==5.4.0
pyzmq